Fix linux compilation

+++ babextended/include/subjectobserver.h
@@ -61,7 +61,7 @@ public:
     
     void notify()
     {
-        for (int i=0; i < mObservers.size(); i++)
+        for (size_t i=0; i < mObservers.size(); i++)
         {
             mObservers[i]->call();
         }
